refactor: extract and move the "case-monitoring" code
Split the code of the main process and the sub process (reduce file size, split responsibilities).
